requirements.txt
black==24.8.0
click==8.1.7
isort==5.13.2
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
python-dotenv==1.0.1
toml==0.10.2
